requirements.txt
# PBIP Studio Requirements
# Python 3.10+ required

# GUI Framework
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
qtawesome>=1.3.0

# Backend API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Visualization (for embedded charts)
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Additional
requests>=2.31.0

# Azure Authentication for Fabric API
azure-identity>=1.15.0
msal>=1.26.0

# Packaging (for building exe/msi)
pyinstaller>=6.3.0
cx-Freeze>=6.15.0

